pyproject.toml
[project]
name = "banking-service"
version = "0.1.0"
description = "Production-ready banking service API with FastAPI and SQLite"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]
e2e = [
    "playwright>=1.45.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S105", "S106", "B008", "S104"]  # S10x: test fixtures, B008: Depends(), S104: 0.0.0.0

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "E501"]  # assert + long lines fine in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
